+++ feature-libs/my-account/organization/src/components/cost-center/budgets/assign/cost-center-assign-budget.service.spec.ts
@@ -1,10 +1,12 @@
+import { Injectable } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { Budget, CostCenterService, EntitiesModel } from '@spartacus/core';
 import {
   IconTestingModule,
   Table,
-  TableTestingModule,
+  TableService,
+  TableStructure,
 } from '@spartacus/storefront';
 import { Observable, of } from 'rxjs';
 import { CostCenterAssignBudgetListService } from './cost-center-assign-budget.service';
@@ -34,18 +36,29 @@ class MockCostCenterService {
   unassignBudget() {}
 }
 
+@Injectable()
+export class MockTableService {
+  buildStructure(type): Observable<TableStructure> {
+    return of({ type });
+  }
+}
+
 describe('CostCenterAssignBudgetListService', () => {
   let service: CostCenterAssignBudgetListService;
   let costCenterService: CostCenterService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [IconTestingModule, RouterTestingModule, TableTestingModule],
+      imports: [IconTestingModule, RouterTestingModule],
       providers: [
         CostCenterAssignBudgetListService,
         {
           provide: CostCenterService,
           useClass: MockCostCenterService,
+        },
+        {
+          provide: TableService,
+          useClass: MockTableService,
         },
       ],
     });
